api/requirements.txt
flask==2.3.3
cassandra-driver==3.29.1